feat: use modules instead of global bindings
