+++ src/components/EscDshotDirection/EscDshotDirectionComponent.js
@@ -5,6 +5,7 @@ import EscDshotDirectionMotorDriver from "./EscDshotDirectionMotorDriver.js";
 import DshotCommand from "../../js/utils/DshotCommand.js";
 import FC from "../../js/fc.js";
 import { API_VERSION_1_44 } from '../../js/data_storage.js';
+import { getMixerImageSrc } from "../../js/utils/common.js";
 
 class EscDshotDirectionComponent
 {
