+++ src/js/Features.js
@@ -1,5 +1,7 @@
 import { bit_check, bit_set, bit_clear } from "./bit";
 import { API_VERSION_1_44 } from './data_storage';
+import semver from "semver";
+import { tracking } from "./Analytics";
 
 const Features = function (config) {
     const self = this;
