+++ src/js/RateCurve.js
@@ -1,5 +1,6 @@
 import FC from "./fc";
 import { API_VERSION_1_43 } from "./data_storage";
+import semver from "semver";
 
 const minRc = 1000;
 const midRc = 1500;
