+++ src/js/TuningSliders.js
@@ -2,6 +2,10 @@ import MSP from "./msp";
 import FC from "./fc";
 import MSPCodes from "./msp/MSPCodes";
 import { API_VERSION_1_44 } from './data_storage';
+import { isExpertModeEnabled } from "./utils/isExportModeEnabled";
+import semver from "semver";
+import { mspHelper } from "./msp/MSPHelper";
+import { TABS } from "./gui";
 
 const TuningSliders = {
     // Legacy Sliders
