+++ src/js/fc.js
@@ -1,5 +1,6 @@
 import { bit_check } from "./bit";
 import { API_VERSION_1_42, API_VERSION_1_43, API_VERSION_1_44, API_VERSION_1_45 } from './data_storage';
+import semver from "semver";
 
 const INITIAL_CONFIG = {
     apiVersion:                       "0.0.0",
