+++ src/js/tabs/adjustments.js
@@ -1,11 +1,12 @@
 import { i18n } from '../localization';
-import GUI from '../gui';
+import GUI, { TABS } from '../gui';
 import { mspHelper } from '../msp/MSPHelper';
 import MSP from '../msp';
 import FC from '../fc';
 import MSPCodes from '../msp/MSPCodes';
 import { API_VERSION_1_42 } from '../data_storage';
 import { gui_log } from '../gui_log';
+import semver from 'semver';
 
 const adjustments = {};
 
@@ -305,7 +306,7 @@ adjustments.adjust_template = function () {
     selectFunction.sortSelect(i18n.getMessage("adjustmentsFunction0"));
 };
 
-window.TABS.adjustments = adjustments;
+TABS.adjustments = adjustments;
 export {
     adjustments,
 };
