+++ src/js/tabs/auxiliary.js
@@ -8,6 +8,7 @@ import MSP from '../msp';
 import MSPCodes from '../msp/MSPCodes';
 import adjustBoxNameIfPeripheralWithModeID from '../peripherals';
 import { gui_log } from '../gui_log';
+import { getTextWidth } from '../utils/common';
 
 const auxiliary = {};
