+++ src/js/tabs/configuration.js
@@ -1,6 +1,6 @@
 import semver from 'semver';
 import { i18n } from '../localization';
-import GUI from '../gui';
+import GUI, { TABS } from '../gui';
 import { tracking } from "../Analytics";
 import { reinitializeConnection } from '../serial_backend';
 import { mspHelper } from '../msp/MSPHelper';
@@ -9,6 +9,7 @@ import MSP from '../msp';
 import MSPCodes from '../msp/MSPCodes';
 import { API_VERSION_1_42, API_VERSION_1_43, API_VERSION_1_45 } from '../data_storage';
 import { gui_log } from '../gui_log';
+import { updateTabList } from '../utils/updateTabList';
 
 const configuration = {
     analyticsChanges: {},
@@ -576,5 +577,5 @@ configuration.cleanup = function (callback) {
     if (callback) callback();
 };
 
-window.TABS.configuration = configuration;
+TABS.configuration = configuration;
 export { configuration };
