+++ src/js/tabs/power.js
@@ -1,11 +1,12 @@
 import { i18n } from '../localization';
-import GUI from '../gui';
+import GUI, { TABS } from '../gui';
 import { tracking } from "../Analytics";
 import { mspHelper } from '../msp/MSPHelper';
 import FC from '../fc';
 import MSP from '../msp';
 import MSPCodes from '../msp/MSPCodes';
 import { gui_log } from '../gui_log';
+import jBox from 'jbox';
 
 const power = {
     supported: false,
@@ -533,5 +534,5 @@ power.cleanup = function (callback) {
     }
 };
 
-window.TABS.power = power;
+TABS.power = power;
 export { power };
