+++ src/js/tabs/sensors.js
@@ -1,5 +1,5 @@
 import { i18n } from "../localization";
-import GUI from '../gui';
+import GUI, { TABS } from '../gui';
 import { get as getConfig, set as setConfig } from '../ConfigStorage';
 import { have_sensor } from "../sensor_helpers";
 import FC from "../fc";
@@ -479,7 +479,7 @@ sensors.cleanup = function (callback) {
     if (callback) callback();
 };
 
-window.TABS.sensors = sensors;
+TABS.sensors = sensors;
 export {
     sensors,
 };
