+++ src/js/utils/EscProtocols.js
@@ -1,3 +1,4 @@
+import semver from 'semver';
 import { API_VERSION_1_42, API_VERSION_1_43 } from '../data_storage';
 
 class EscProtocols
