+++ src/js/utils/isExportModeEnabled.js
@@ -1,0 +1,3 @@
+export function isExpertModeEnabled() {
+    return $('input[name="expertModeCheckbox"]').is(':checked');
+}
