+++ src/js/utils/showErrorDialog.js
@@ -1,0 +1,12 @@
+
+export function showErrorDialog(message) {
+   const dialog = $('.dialogError')[0];
+
+    $('.dialogError-content').html(message);
+
+    $('.dialogError-closebtn').click(function() {
+        dialog.close();
+    });
+
+    dialog.showModal();
+}
