+++ src/js/utils/updateTabList.js
@@ -1,0 +1,53 @@
+import semver from "semver";
+import { API_VERSION_1_42 } from "../data_storage";
+import FC from "../fc";
+import { isExpertModeEnabled } from "./isExportModeEnabled";
+
+export function updateTabList(features) {
+
+    if (isExpertModeEnabled()) {
+        $('#tabs ul.mode-connected li.tab_failsafe').show();
+        $('#tabs ul.mode-connected li.tab_adjustments').show();
+        $('#tabs ul.mode-connected li.tab_servos').show();
+        $('#tabs ul.mode-connected li.tab_sensors').show();
+        $('#tabs ul.mode-connected li.tab_logging').show();
+    } else {
+        $('#tabs ul.mode-connected li.tab_failsafe').hide();
+        $('#tabs ul.mode-connected li.tab_adjustments').hide();
+        $('#tabs ul.mode-connected li.tab_servos').hide();
+        $('#tabs ul.mode-connected li.tab_sensors').hide();
+        $('#tabs ul.mode-connected li.tab_logging').hide();
+    }
+
+    if (features.isEnabled('GPS') && isExpertModeEnabled()) {
+        $('#tabs ul.mode-connected li.tab_gps').show();
+    } else {
+        $('#tabs ul.mode-connected li.tab_gps').hide();
+    }
+
+    if (features.isEnabled('LED_STRIP')) {
+        $('#tabs ul.mode-connected li.tab_led_strip').show();
+    } else {
+        $('#tabs ul.mode-connected li.tab_led_strip').hide();
+    }
+
+    if (features.isEnabled('TRANSPONDER')) {
+        $('#tabs ul.mode-connected li.tab_transponder').show();
+    } else {
+        $('#tabs ul.mode-connected li.tab_transponder').hide();
+    }
+
+    if (features.isEnabled('OSD')) {
+        $('#tabs ul.mode-connected li.tab_osd').show();
+    } else {
+        $('#tabs ul.mode-connected li.tab_osd').hide();
+    }
+
+    $('#tabs ul.mode-connected li.tab_power').show();
+
+    if (semver.gte(FC.CONFIG.apiVersion, API_VERSION_1_42)) {
+        $('#tabs ul.mode-connected li.tab_vtx').show();
+    } else {
+        $('#tabs ul.mode-connected li.tab_vtx').hide();
+    }
+}
