+++ src/main.html
@@ -58,30 +58,21 @@
     <!-- CORDOVA_INCLUDE js/cordova_chromeapi.js -->
     <!-- CORDOVA_INCLUDE js/cordova_startup.js -->
     <script type="text/javascript" src="./node_modules/lru_map/lru.js"></script>
-    <script type="text/javascript" src="./node_modules/marked/marked.min.js"></script>
-    <script type="text/javascript" src="./node_modules/dompurify/dist/purify.min.js"></script>
-    <script type="text/javascript" src="./node_modules/universal-ga/lib/analytics.min.js"></script>
-    <script type="text/javascript" src="./node_modules/short-unique-id/dist/short-unique-id.min.js"></script>
-    <script type="text/javascript" src="./node_modules/object-hash/dist/object_hash.js"></script>
     <script type="text/javascript" src="./node_modules/jquery/dist/jquery.min.js"></script>
-    <script type="text/javascript" src="./node_modules/jbox/dist/jBox.min.js"></script>
     <script type="text/javascript" src="./node_modules/jquery-ui-npm/jquery-ui.min.js"></script>
-    <script type="text/javascript" src="./node_modules/djv/djv.js"></script>
     <script type="text/javascript" src="./js/libraries/d3.min.js"></script>
     <script type="text/javascript" src="./js/libraries/jquery.nouislider.all.min.js"></script>
     <script type="text/javascript" src="./node_modules/three/build/three.min.js"></script>
     <script type="text/javascript" src="./node_modules/three/examples/js/renderers/CanvasRenderer.js"></script>
     <script type="text/javascript" src="./node_modules/three/examples/js/renderers/Projector.js"></script>
     <script type="text/javascript" src="./js/libraries/jquery.flightindicators.js"></script>
-    <script type="text/javascript" src="./node_modules/semver-min/umd/index.js"></script>
-    <script type="text/javascript" src="./node_modules/switchery-latest/dist/switchery.min.js"></script>
     <script type="text/javascript" src="./node_modules/bluebird/js/browser/bluebird.min.js"></script>
     <script type="text/javascript" src="./js/libraries/jquery.ba-throttle-debounce.min.js"></script>
-    <script type="text/javascript" src="./node_modules/inflection/lib/inflection.js"></script>
     <script type="text/javascript" src="./node_modules/jquery-textcomplete/dist/jquery.textcomplete.min.js"></script>
     <script type="text/javascript" src="./node_modules/jquery-touchswipe/jquery.touchSwipe.min.js"></script>
     <script type="text/javascript" src="./node_modules/select2/dist/js/select2.min.js"></script>
     <script type="text/javascript" src="./node_modules/multiple-select/dist/multiple-select.min.js"></script>
+    <script type="text/javascript" src="./node_modules/object-hash/dist/object_hash.js"></script>
     <script type="module" src="./js/main.js"></script>
 
     <title></title>
